+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Removed CocoaPods support.
 - Moved SwiftLint to use the system version.
 - Fixed an issue with the HUD demo, where the center image was not being displayed.
+- Improved the way that images are drawn.
 
 ## 2.5.1